Rename the balance calculation function
In the past the balance of a user would be calculated when needed.
But with a growing number of transaction this was a bottleneck.
The function will rename with its new name to allow periodic checks
to confirm the new algorithm works as expected.

This change is in reference to US#56 and task #57