Cascade transaction deletions
Currently the only way to delete transactions is the "clear" endpoint,
which is only used during testing and deactivated in production.
This is by design, as deletion of transactions should be impossible.
But in when deleting a transaction, it's pair transaction also must be
deleted to satisfy the double entry accounting constraint.